Split out gmvsmelee (and 2 others) from start of gm_1A4C
This filename is known from an assert string. The file handles Major
Scene ID 2 (VS Melee) and its associated major scene callbacks.

Other files split out have made up names:

* gmmenu: contains the OnLoad callback for scene ID 1 (main menu), but
  contains some other OnLoad callbacks so may need to be further split

* gmscdata: contains the large MajorScene and MinorSceneHandler arrays,
  which are provided to other files solely via getter function helpers
  and can thus be declared statically in this file
