+++ configure.py
@@ -861,6 +861,9 @@ config.libs = [
             Object(NonMatching, "melee/gm/gm_1A36.c"),
             Object(NonMatching, "melee/gm/gm_1A3F.c"),
             Object(NonMatching, "melee/gm/gm_1A45.c"),
+            Object(NonMatching, "melee/gm/gmscdata.c"),
+            Object(NonMatching, "melee/gm/gmmenu.c"),
+            Object(NonMatching, "melee/gm/gmvsmelee.c"),
             Object(NonMatching, "melee/gm/gm_1A4C.c"),
             Object(Matching, "melee/gm/gm_1AED.c"),
             Object(NonMatching, "melee/gm/gm_1B03.c"),
