+++ src/melee/gm/gm_1A3F.c
@@ -6,6 +6,7 @@
 #include "gmmain_lib.h"
 
 #include "db/db.h"
+#include "gm/gmscdata.h"
 #include "lb/lbaudio_ax.h"
 #include "lb/lbdvd.h"
 #include "lb/lbheap.h"
