+++ src/melee/gm/gm_1A4C.h
@@ -5,31 +5,6 @@
 #include <melee/gm/forward.h>
 #include <baselib/forward.h>
 
-/* 1A50A0 */ MinorSceneHandler* gm_801A50A0(void);
-/* 1A50AC */ MajorScene* gm_801A50AC(void);
-/* 1A50B8 */ UNK_RET gm_801A50B8_OnLoad(UNK_PARAMS);
-/* 1A5130 */ UNK_RET gm_801A5130_OnLoad(UNK_PARAMS);
-/* 1A51A8 */ UNK_RET gm_801A51A8_OnLoad(UNK_PARAMS);
-/* 1A5220 */ UNK_RET gm_801A5220_OnLoad(UNK_PARAMS);
-/* 1A5244 */ s8* gm_801A5244(void);
-/* 1A5250 */ UNK_T gm_801A5250(void);
-/* 1A5258 */ UNK_RET gm_801A5258(UNK_PARAMS);
-/* 1A52D0 */ UNK_RET gm_801A52D0(UNK_PARAMS);
-/* 1A5360 */ UNK_RET gm_801A5360(UNK_PARAMS);
-/* 1A5598 */ UNK_RET gm_801A5598_OnInit(UNK_PARAMS);
-/* 1A55C4 */ UNK_RET gm_801A55C4(UNK_PARAMS);
-/* 1A55EC */ UNK_RET gm_801A55EC_OnLoad(UNK_PARAMS);
-/* 1A5614 */ void gm_801A5614_OnUnload(void);
-/* 1A5618 */ UNK_RET gm_801A5618(UNK_PARAMS);
-/* 1A5680 */ UNK_RET gm_801A5680(UNK_PARAMS);
-/* 1A5754 */ UNK_RET gm_801A5754(UNK_PARAMS);
-/* 1A57A8 */ UNK_RET gm_801A57A8(UNK_PARAMS);
-/* 1A583C */ UNK_RET gm_801A583C(UNK_PARAMS);
-/* 1A5AF0 */ UNK_RET gm_801A5AF0(UNK_PARAMS);
-/* 1A5C3C */ UNK_RET gm_801A5C3C(UNK_PARAMS);
-/* 1A5EC8 */ UNK_RET gm_801A5EC8(UNK_PARAMS);
-/* 1A5F00 */ UNK_RET gm_801A5F00(UNK_PARAMS);
-/* 1A5F64 */ UNK_RET gm_801A5F64(UNK_PARAMS);
 /* 1A6254 */ UNK_RET gm_801A6254(UNK_PARAMS);
 /* 1A6308 */ UNK_RET gm_801A6308(UNK_PARAMS);
 /* 1A632C */ UNK_RET gm_801A632C_OnEnter(UNK_PARAMS);
