+++ src/melee/gm/gmmenu.c
@@ -1,0 +1,11 @@
+#include "gmmenu.h"
+
+#include <melee/gm/types.h>
+
+/// #gm_801A50B8_OnLoad
+
+/// #gm_801A5130_OnLoad
+
+/// #gm_801A51A8_OnLoad
+
+/// #gm_801A5220_OnLoad
