+++ src/melee/gm/gmscdata.h
@@ -1,0 +1,9 @@
+#ifndef MELEE_GM_SCDATA_H
+#define MELEE_GM_SCDATA_H
+
+#include <melee/gm/forward.h>
+
+MinorSceneHandler* gm_801A50A0(void);
+MajorScene* gm_801A50AC(void);
+
+#endif
