+++ src/melee/gm/gmvsmelee.c
@@ -1,0 +1,50 @@
+#include "gmvsmelee.h"
+
+#include <melee/gm/types.h>
+#include <melee/gm/gmmain_lib.h>
+
+/* 4D6730 */ extern UNK_T gm_804D6730;
+
+s8* gm_801A5244(void)
+{
+    return &gmMainLib_804D3EE0->unk_590;
+}
+
+UNK_T gm_801A5250(void)
+{
+    return &gm_804D6730;
+}
+
+/// #gm_801A5258
+
+/// #gm_801A52D0
+
+/// #gm_801A5360
+
+/// #gm_801A5598_OnInit
+
+/// #gm_801A55C4
+
+/// #gm_801A55EC_OnLoad
+
+void gm_801A5614_OnUnload(void) {}
+
+/// #gm_801A5618
+
+/// #gm_801A5680
+
+/// #gm_801A5754
+
+/// #gm_801A57A8
+
+/// #gm_801A583C
+
+/// #gm_801A5AF0
+
+/// #gm_801A5C3C
+
+/// #gm_801A5EC8
+
+/// #gm_801A5F00
+
+/// #gm_801A5F64
